move circle tool to utilities
