+++ buildings/utilities/circle_tool.py
